requirements.txt
Flask==2.1.2
numpy==1.22.3
Pillow==9.1.0
tensorflow-cpu==2.8.0
tensorflow-estimator==2.8.0
gunicorn==20.0.4